feat(Borrowed/Rate/Tvl): consolidate replicated code with base repositories
